Improve extractors organization
