+++ src/IReplacementDefinitionsExtractor.cs
