+++ tests/Extractors/CsReplacementDefinitionsExtractorTests.cs
@@ -2,7 +2,7 @@
 using NUnit.Framework;
 using RDumont.NugetContentGenerator.Runtime.Extractors;
 
-namespace RDumont.NugetContentGenerator.Runtime.Tests
+namespace RDumont.NugetContentGenerator.Runtime.Tests.Extractors
 {
     [TestFixture]
     public class CsReplacementDefinitionsExtractorTests
@@ -126,6 +126,24 @@ contents";
 
             // Assert
             Assert.That(isEndOfBlock, Is.True);
+        }
+    }
+
+    public class TestableCsExtractor : CsReplacementDefinitionsExtractor
+    {
+        public new string GetReplacementLine(string line)
+        {
+            return base.GetReplacementLine(line);
+        }
+
+        public new bool IsStartOfBlock(string line)
+        {
+            return base.IsStartOfBlock(line);
+        }
+
+        public new bool IsEndOfBlock(string line)
+        {
+            return base.IsEndOfBlock(line);
         }
     }
 }
